Move test back to the websocket module
